Move merge_replacements to PropagatorState.
