+++ angr/analyses/propagator/propagator.py
@@ -281,7 +281,7 @@ class PropagatorAnalysis(ForwardAnalysis, Analysis):  # pylint:disable=abstract-
         if self.model.replacements is None:
             self.model.replacements = state._replacements
         else:
-            self.merge_replacements(self.model.replacements, state._replacements)
+            PropagatorState.merge_replacements(self.model.replacements, state._replacements)
 
         self.model.equivalence |= state._equivalence
 
@@ -377,44 +377,6 @@ class PropagatorAnalysis(ForwardAnalysis, Analysis):  # pylint:disable=abstract-
             self._analysis_core_graph()
 
         self._post_analysis()
-
-    @staticmethod
-    def merge_replacements(replacements_0, replacements_1) -> bool:
-        """
-        The replacement merging logic is special: replacements_1 is the newer replacement result and replacement_0 is
-        the older result waiting to be updated. When both replacements_1 and replacement_0 have a non-top value for the
-        same variable and code location, we will update the slot in replacement_0 with the value from replacement_1.
-
-        :return:            Whether merging has happened or not.
-        """
-        merge_occurred = False
-        for loc, vars_ in replacements_1.items():
-            if loc not in replacements_0:
-                replacements_0[loc] = vars_.copy()
-                merge_occurred = True
-            else:
-                for var, repl in vars_.items():
-                    if var not in replacements_0[loc]:
-                        replacements_0[loc][var] = repl
-                        merge_occurred = True
-                    else:
-                        if PropagatorState.is_top(repl) or PropagatorState.is_top(replacements_0[loc][var]):
-                            t = PropagatorState.top(repl.bits if isinstance(repl, ailment.Expression) else repl.size())
-                            replacements_0[loc][var] = t
-                            merge_occurred = True
-                        elif (
-                            isinstance(replacements_0[loc][var], claripy.ast.Base) or isinstance(repl, claripy.ast.Base)
-                        ) and replacements_0[loc][var] is not repl:
-                            replacements_0[loc][var] = repl
-                            merge_occurred = True
-                        elif (
-                            not isinstance(replacements_0[loc][var], claripy.ast.Base)
-                            and not isinstance(repl, claripy.ast.Base)
-                            and replacements_0[loc][var] != repl
-                        ):
-                            replacements_0[loc][var] = repl
-                            merge_occurred = True
-        return merge_occurred
 
 
 register_analysis(PropagatorAnalysis, "Propagator")
